確定予想を締切 2〜4分前 に移した（下落の7割は8分前→直前で起きていた）
lab winner-drift の実測（当たった目176件・同じレースで比較）:
  8分前のオッズ起点  下落 -13.8% → ×1.5 に必要な倍率 ×1.74
  直前(2〜4分前)起点 下落  -4.2% → 同 ×1.57
時刻を寄せるだけで下落の約7割が消え、的中率は1ptも払わない。必要倍率が
×1.57 まで下がるので、いまの ×1.62 のままで中央値を守れる。

倍率で対応する道が割に合わないことも確定した。本命10点の当たり目が乗る帯
（2〜30倍・89件）で「当たって元本割れ」は ×1.50 22% / ×1.62 18% /
×1.76 13% / ×2.00 7%。×2.00 まで上げても残るので、締切前オッズで組む限り
「当たってマイナスにならない」は原理的に保証できない。

実装:
- FINAL_MIN, FINAL_MAX = 2, 4（旧 4, 10）。スケジューラは30秒ごとに回るので
  2分幅でも4回試行できる。同じループで直前に3連単を取り直しており、予想は
  そのオッズを使う。
- 取りこぼし救済 FINAL_RESCUE_MIN=0.7。窓が狭いぶん、締切0.7〜2分前に
  未作成のレースを拾う（「どのレースにも確定予想がある」を壊さないため）。
- record_late_modes の呼び出しを外した（二重になるため。関数とテストは残す）
- predict_pending の分数引数を int → float
- winner-drift の「8分前 vs 直前」節は hon_late 依存をやめ、オッズの
  スナップショット同士で比べる形に変更（時刻変更をまたいでも同じ物差し）

倍率は ×1.62 のまま据え置き。9/16 を境に条件が変わるので前後を混ぜない
（post_time_at_pred − created_at で判別できる）。

出荷前チェック: 46 tests（窓とfloat引数を固定する2件を追加）。合成データで
スナップショット比較が -20%→-5.9% を復元することを確認。

+++ boatlab/ops/daily.py
@@ -184,7 +184,7 @@ def staking_from_settings(row: SettingsVersion) -> StakingParams:
 
 
 def predict_pending(predictor: Predictor, stage: str, role: str = "active", d: date | None = None,
-                    min_minutes_before_close: int = 4, max_minutes_before_close: int | None = None,
+                    min_minutes_before_close: float = 4, max_minutes_before_close: float | None = None,
                     now: datetime | None = None, hist_cache: HistoryFrames | None = None) -> dict:
     """締切まで min_minutes 以上あるレースに予想を保存する（同一 stage/role は1回のみ）。
